refactor/EV-872/Overall refactor on the project folders
