+++ lib/modules/wallet/widgets/crypto_list_view.dart
@@ -6,7 +6,7 @@ import '../../../shared/widgets/default_error_widget.dart';
 import '../../../shared/widgets/default_loading_spinner.dart';
 import '../model/crypto_data_view_data.dart';
 import '../model/user_wallet_model.dart';
-import '../providers/providers.dart';
+import '../controllers/providers.dart';
 import 'crypto_list_tile.dart';
 
 class CryptoListView extends HookConsumerWidget {
